Move contact data into separate module
